Finnhub_Producer/requirements.txt
kafka-python
websocket-client
finnhub-python